Added test to functional.rs and slothash code within the processor.rs function
